Leverage the Netflix Domain Graph Service (DGS) Framework for managing GraphQL
This provides a better integration with Spring Boot.
