+++ application/build.gradle
@@ -54,10 +54,12 @@ dependencies {
     implementation "com.fasterxml.jackson.module:jackson-module-kotlin:${jacksonVersion}"
     implementation "com.fasterxml.jackson.dataformat:jackson-dataformat-csv:${jacksonVersion}"
     runtimeOnly('org.springframework.boot:spring-boot-devtools')
-    implementation "com.graphql-java-kickstart:graphql-spring-boot-starter:${graphqlSpringBootStarterVersion}"
-    // to embed GraphiQL tool
-    implementation "com.graphql-java-kickstart:graphiql-spring-boot-starter:${graphqlSpringBootStarterVersion}"
-    implementation "com.graphql-java-kickstart:graphql-java-tools:11.0.1"
+
+    // GraphQL and GraphiQL (via Netflix Data Graph Service (DGS))
+    implementation(platform('com.netflix.graphql.dgs:graphql-dgs-platform-dependencies:4.9.11'))
+    implementation "com.netflix.graphql.dgs:graphql-dgs-spring-boot-starter"
+    implementation "com.graphql-java:graphql-java-extended-scalars:17.0"
+
     testImplementation('org.springframework.boot:spring-boot-starter-test')
     testImplementation('org.springframework.security:spring-security-test')
 }
